- synced with SVN

+++ DependentExtensions/Swig/MakeSwig.bat
@@ -1,7 +1,8 @@
+@echo off
+
 REM This file was taken from RakNet 4.082 without any modifications.
 REM Please see licenses/RakNet license.txt for the underlying license and related copyright.
 
-@echo off
 if "%1"=="" goto :NOVARS
 echo Performing Swig build
 set swigPath=%2
@@ -9,13 +10,15 @@ if "%swigPath%"=="" goto :SKIPADDSLASH
 if "%swigPath:~-1%"=="\" goto :SKIPADDSLASH 
 SET swigPath=%swigPath%\
 :SKIPADDSLASH
-del /F /Q SwigOutput\SwigCSharpOutput\*
+if exist SwigOutput\SwigCSharpOutput del /F /Q SwigOutput\SwigCSharpOutput\*
 if "%3"=="" goto :NOSQL
 %swigPath%swig -c++ -csharp -namespace RakNet -I"%1" -I"SwigInterfaceFiles" -I"%3" -DSWIG_ADDITIONAL_SQL_LITE -outdir SwigOutput\SwigCSharpOutput -o SwigOutput\CplusDLLIncludes\RakNet_wrap.cxx SwigInterfaceFiles\RakNet.i
+if errorlevel 1 GOTO :SWIGERROR
 copy /Y SwigOutput\SwigCSharpOutput\* SwigWindowsCSharpSample\SwigTestApp\SwigFiles\*
 GOTO ENDSWIG
 :NOSQL
 %swigPath%swig -c++ -csharp -namespace RakNet -I"%1" -I"SwigInterfaceFiles" -outdir SwigOutput\SwigCSharpOutput -o SwigOutput\CplusDLLIncludes\RakNet_wrap.cxx SwigInterfaceFiles\RakNet.i
+if errorlevel 1 GOTO :SWIGERROR
 copy /Y SwigOutput\SwigCSharpOutput\* SwigWindowsCSharpSample\SwigTestApp\SwigFiles\*
 :ENDSWIG
 if errorlevel 1 GOTO :SWIGERROR
